Document extra settings for the capistrano-rbenv plugin
I found out the extra settings that capistrano-rbenv needs to be able to detect the small differences with the "regular" rbenv. I've added to the docs in case someone else needs it to. Let me know if you think its ok, or if it needs more explanation or changes.

Thanks!