CI/CD: fix Bintray production publishing
We used to publish all packages to a Bintray package version named 'all'.
But it turns out that 1 year after its creation, Bintray disallows
uploading new files to that Bintray package version.

So now we create a new Bintray package version every time we upload a
new set of packages.

Closes #60.