Fix fullstaq-rbenv package version number
The version number embedded in the RPM/DEB file did not contain the a version number suffix. Instead of for example "1.1.2-16", it would contain "1.1.2".

This is because the rbenv package building script tries to infer the package number by itself by running `bin/rbenv` inside the Docker container. But that container has no Git installed, so it's not able to infer the suffix.

We fix this by requiring the user to specify the full version number, including suffix.

We need to yank a bunch of packages because of this.