CI/CD: fix publishing architecture-independent Debian packages in combination with introducing a new distro
If an architecture-independent package will be imported into at least one repo, then
ensure that they'll be imported into all distros.

This is because Aptly performs deduplication. When an architecture-independent repo
is imported into some repos but not all of them, then the repos for which it's not
imported will have outdated metadata.

https://github.com/fullstaq-labs/fullstaq-ruby-server-edition/pull/85#issuecomment-940273331