Use sccache instead of ccache (#99)
Closes #86.

This requires https://github.com/mozilla/sccache/pull/1109, which has been merged but hasn't been released yet. Therefore we use our own sccache binary which includes this feature. We'll use the upstream binary once sccache 0.3.0 is out.

In the Ubuntu 20.04 image, this undoes a fix for #80 because that fix never worked in the first place. See #101.